implement getoptionvalue
