+++ .vscode/settings.json
@@ -1,6 +1,12 @@
 {
   "cSpell.words": [
     "approvable",
+<<<<<<< Updated upstream
+=======
+    "cmdk",
+    "Fjxj",
+    "JEMM",
+>>>>>>> Stashed changes
     "mindee"
   ]
 }
