+++ lib/utils.ts
@@ -35,3 +35,40 @@ export const sliceWithEllipsis = (
     ? str.slice(0, length) + '...' + (suffix ? suffix : '')
     : str;
 };
+
+export function stringSimilarity(str1: string, str2: string): number {
+  const distance = levenshtein(str1, str2);
+  return 1 - distance / Math.max(str1.length, str2.length);
+}
+
+function levenshtein(a: string, b: string): number {
+  const an = a ? a.length : 0;
+  const bn = b ? b.length : 0;
+  if (an === 0) {
+    return bn;
+  }
+  if (bn === 0) {
+    return an;
+  }
+  const matrix = Array(an + 1);
+  for (let i = 0; i <= an; i++) {
+    matrix[i] = Array(bn + 1).fill(0);
+    matrix[i][0] = i;
+  }
+  for (let j = 1; j <= bn; j++) {
+    matrix[0][j] = j;
+  }
+  for (let i = 1; i <= an; i++) {
+    for (let j = 1; j <= bn; j++) {
+      if (a[i - 1] === b[j - 1]) {
+        matrix[i][j] = matrix[i - 1][j - 1];
+      } else {
+        matrix[i][j] = Math.min(
+          matrix[i - 1][j - 1] + 1,
+          Math.min(matrix[i][j - 1] + 1, matrix[i - 1][j] + 1),
+        );
+      }
+    }
+  }
+  return matrix[an][bn];
+}
